Improve system integrity
