+++ bb/src/bbserver.cc
@@ -142,27 +142,6 @@ void switchIds(txp::Msg* pMsg)
     return;
 }
 
-void switchIdsToMountPoint(txp::Msg* pMsg)
-{
-    const uid_t l_Owner = (uid_t)((txp::Attr_uint32*)pMsg->retrieveAttrs()->at(txp::mntptuid))->getData();
-    const gid_t l_Group = (gid_t)((txp::Attr_uint32*)pMsg->retrieveAttrs()->at(txp::mntptgid))->getData();
-
-    int rc = becomeUser(l_Owner, l_Group);
-    if (!rc)
-    {
-        bberror << err("in.misc.mntptuid", l_Owner) << err("in.misc.mntptgid", l_Group);
-    }
-    else
-    {
-        stringstream errorText;
-        errorText << "becomeUser failed";
-    	bberror << err("error.uid", l_Owner) << err("error.gid", l_Group);
-        LOG_ERROR_TEXT_RC_AND_BAIL(errorText, rc);
-    }
-
-    return;
-}
-
 
 //*****************************************************************************
 //  Requests from bbproxy
@@ -524,7 +503,7 @@ void msgin_createlogicalvolume(txp::Id id, const std::string& pConnectionName, t
         switchIdsToMountPoint(msg);
 
         // NOTE:  If addLogicalVolume() fails, it fills in the error state...
-        rc = addLogicalVolume(pConnectionName, l_HostName, l_LVKeyPtr, l_JobId, (TOLERATE_ALREADY_EXISTS_OPTION)l_Option);
+        rc = addLogicalVolume(pConnectionName, l_HostName, msg, l_LVKeyPtr, l_JobId, (TOLERATE_ALREADY_EXISTS_OPTION)l_Option);
         if (rc)
         {
             if (rc < 0)
